+++ assembly/src/library/mod.rs
@@ -101,6 +101,16 @@ impl Library {
             ..self
         }
     }
+
+    /// Clears all debug information in order to reduce the library size.
+    pub fn clear_debug_info(self) -> Self {
+        let mut mast_forest = (*self.mast_forest).clone();
+        mast_forest.clear_debug_info();
+        Self {
+            mast_forest: Arc::new(mast_forest),
+            ..self
+        }
+    }
 }
 
 // ------------------------------------------------------------------------------------------------