+++ assembly/src/tests.rs
@@ -3302,3 +3302,43 @@ fn vendoring() -> TestResult {
 fn test_assert_diagnostic_lines() {
     assert_diagnostic_lines!(report!("the error string"), "the error string", "other", "lines");
 }
+
+#[test]
+fn test_clear_debug_info() {
+    let source = "
+    const.ERR1=\"oh no\"
+
+    const.DEFAULT_CONST=100
+
+    proc.foo
+        push.1.2 add
+        debug.stack.8
+    end
+
+    begin
+        emit.DEFAULT_CONST
+
+        exec.foo
+
+        debug.stack.4
+
+        drop
+
+        trace.DEFAULT_CONST
+
+        assert.err=ERR1
+    end
+    ";
+
+    let assembler = Assembler::default().with_debug_mode(true);
+    let program = assembler.assemble_program(source).unwrap();
+
+    let mut target_with_debug_info = Vec::new();
+    program.write_into(&mut target_with_debug_info);
+
+    let mut target_without_debug_info = Vec::new();
+    program.clear_debug_info().write_into(&mut target_without_debug_info);
+
+    assert!(target_with_debug_info.len() == 588);
+    assert!(target_without_debug_info.len() == 158);
+}