+++ core/src/mast/node/join_node.rs
@@ -107,6 +107,12 @@ impl JoinNode {
     pub fn after_exit(&self) -> &[DecoratorId] {
         &self.after_exit
     }
+
+    /// Clears the decorators.
+    pub fn clear_decorators(&mut self) {
+        self.before_enter.clear();
+        self.after_exit.clear();
+    }
 }
 
 /// Mutators